Move player and team profile routes to dedicated blueprints
Extracts ten player routes (list, GET/PUT, profile-photo POST/DELETE,
and five injury CRUD handlers) into a new app/routes/players.py, and
six team routes (list, GET/PUT, players list, profile-photo POST/DELETE)
into a new app/routes/teams.py. Both blueprints register at /_api;
URL surface is unchanged.

Adds a small app/utils/profile_photo_helpers.py module so the two
shared profile-photo helpers don't need to be cross-imported between
the new route files. _injury_json moves with the injury routes as a
private helper inside players.py. _dt_iso is imported transitionally
from _api.py and will move to app/utils/datetime_helpers.py in a
later change.
