+++ app/__init__.py
@@ -243,6 +243,8 @@ def create_app(config: dict | None = None) -> Flask:
     from app.routes._api import bp as _api_bp
     from app.routes.content import bp as content_bp
     from app.routes.penalty_types import bp as penalty_types_bp
+    from app.routes.players import bp as players_bp
+    from app.routes.teams import bp as teams_bp
 
     app.register_blueprint(_api_bp)
     app.register_blueprint(auth_bp)
@@ -253,6 +255,8 @@ def create_app(config: dict | None = None) -> Flask:
     app.register_blueprint(sidecomps_bp)
     app.register_blueprint(content_bp)
     app.register_blueprint(penalty_types_bp)
+    app.register_blueprint(players_bp)
+    app.register_blueprint(teams_bp)
 
     # Register template filters
     from app import filters
