+++ app/routes/_api.py
@@ -11,8 +11,7 @@ from datetime import datetime, timezone
 import collections
 import os
 from flask_login import current_user, login_required
-from sqlalchemy import or_, and_, func
-from sqlalchemy.orm import joinedload
+from sqlalchemy import func
 from sqlalchemy.orm.attributes import flag_modified
 from app.services._common import current_user_type, Scope
 from app.services.permission_service import PermissionService
@@ -1931,7 +1930,6 @@ def tournament_bracket_upload_bytes_api(tournament_url):
     db.session.remove()
 
     from flask import current_app
-    import os
     from datetime import datetime, timezone
 
     original_name = request.args.get("filename", "bracket.png")
@@ -3202,613 +3200,6 @@ def tournament_match_state(tournament_url):
     )
 
 
-@bp.route("/players", methods=["GET"])
-def players_list():
-    """List players with optional search and pagination."""
-    search = request.args.get("search", "").strip()
-    page = max(1, request.args.get("page", type=int) or 1)
-    per_page = 50
-    if search:
-        q = Player.query.filter(Player.name.contains(search) | Player.id.contains(search))
-    else:
-        q = Player.query
-    total = q.count()
-    total_pages = (total + per_page - 1) // per_page
-    players = q.order_by(Player.name.asc()).offset((page - 1) * per_page).limit(per_page).all()
-    return jsonify(
-        {
-            "players": [
-                {
-                    "id": p.id,
-                    "name": p.name,
-                    "profile_photo": p.profile_photo,
-                    "location": p.location,
-                }
-                for p in players
-            ],
-            "page": page,
-            "total_pages": total_pages,
-            "total": total,
-        }
-    )
-
-
-@bp.route("/players/<player_id>", methods=["GET"])
-def player_profile(player_id):
-    """Player profile (public)."""
-    player = Player.query.get(player_id)
-    if not player:
-        return jsonify({"error": "Not found"}), 404
-    regs = PlayerRegistration.query.filter_by(player=player_id).all()
-    can_see_private = current_user.is_authenticated and current_user.id == player_id
-    injuries_query = Injury.query.filter_by(player=player_id)
-    if not can_see_private:
-        injuries_query = injuries_query.filter_by(show=True)
-    injuries = injuries_query.order_by(Injury.stamp.desc()).all()
-    player_notes = []
-    if current_user.is_authenticated:
-        try:
-            all_player_notes = (
-                MatchNote.query.filter_by(player_id=player_id)
-                .options(joinedload(MatchNote.penalty_type))
-                .order_by(MatchNote.created_at.desc())
-                .all()
-            )
-            _all_note_match_ids = [n.match for n in all_player_notes if n.match]
-            _all_matches_by_id = (
-                {m.uuid: m for m in Match.query.filter(Match.uuid.in_(_all_note_match_ids)).all()}
-                if _all_note_match_ids
-                else {}
-            )
-            for note in all_player_notes:
-                can_see_note = False
-                if current_user.id == player_id:
-                    can_see_note = True
-                elif is_player(current_user):
-                    match_obj = _all_matches_by_id.get(note.match) if note.match else None
-                    if match_obj and can_head_ref_match(match_obj.event, current_user.id, match=match_obj):
-                        can_see_note = True
-                if can_see_note:
-                    player_notes.append(note)
-        except Exception:
-            player_notes = []
-
-    note_match_ids = [n.match for n in player_notes if n.match]
-    matches_by_id = (
-        {m.uuid: m for m in Match.query.filter(Match.uuid.in_(note_match_ids)).all()} if note_match_ids else {}
-    )
-
-    penalty_type_ids = {
-        getattr(n, "penalty_type_id", None) for n in player_notes if getattr(n, "penalty_type_id", None)
-    }
-    pt_map = {}
-    if penalty_type_ids:
-        for pt in PenaltyType.query.filter(PenaltyType.id.in_(penalty_type_ids)).all():
-            pt_map[pt.id] = {"name": pt.name, "color": pt.color, "desc": pt.desc or ""}
-
-    player_note_rows = []
-    if player_notes:
-        match_to_points = {}
-        for note in player_notes:
-            idx = "-"
-            match_obj = matches_by_id.get(note.match) if note.match else None
-            if match_obj and note.point_id:
-                match_id = match_obj.uuid
-                if match_id not in match_to_points:
-                    pts = Point.query.filter_by(match=match_id).order_by(Point.stamp).all()
-                    match_to_points[match_id] = [p.uuid for p in pts]
-                order = match_to_points.get(match_id, [])
-                if note.point_id in order:
-                    idx = order.index(note.point_id) + 1
-            pt_id = getattr(note, "penalty_type_id", None)
-            pt_rel = getattr(note, "penalty_type", None)
-            if pt_rel is not None:
-                pt_info = {
-                    "name": pt_rel.name,
-                    "color": pt_rel.color,
-                    "desc": pt_rel.desc or "",
-                }
-            else:
-                pt_info = pt_map.get(pt_id) if pt_id else None
-            if pt_info is None and pt_id:
-                _pt = PenaltyType.query.get(pt_id)
-                if _pt:
-                    pt_info = {
-                        "name": _pt.name,
-                        "color": _pt.color,
-                        "desc": _pt.desc or "",
-                    }
-            player_note_rows.append(
-                {
-                    "created_at": _dt_iso(note.created_at),
-                    "text": note.text or "",
-                    "point_index": str(idx),
-                    "penalty_type_id": pt_id,
-                    "penalty_type_name": pt_info["name"] if pt_info else None,
-                    "penalty_type_color": pt_info["color"] if pt_info else None,
-                    "penalty_type_desc": pt_info.get("desc", "") if pt_info else None,
-                    "match": (
-                        {
-                            "event": match_obj.event if match_obj else None,
-                            "uuid": match_obj.uuid if match_obj else None,
-                            "name": match_obj.name if match_obj else None,
-                        }
-                        if match_obj
-                        else None
-                    ),
-                }
-            )
-
-    event_urls = {r.event for r in regs if r.event}
-    league_ids = {r.league_id for r in regs if r.league_id and not r.event}
-    tournaments_by_url = (
-        {t.url: t for t in Tournament.query.filter(Tournament.url.in_(event_urls)).all()} if event_urls else {}
-    )
-    leagues_by_id = {lg.url: lg for lg in League.query.filter(League.url.in_(league_ids)).all()} if league_ids else {}
-
-    team_keys_needed = set()
-    for r in regs:
-        if not r.team:
-            continue
-        if r.event:
-            team_keys_needed.add(("event", r.event, r.team))
-        elif r.league_id:
-            team_keys_needed.add(("league", r.league_id, r.team))
-
-    team_pseudonym_by_key = {}
-    team_shortname_by_key = {}
-    event_team_pairs = [(e, t) for kind, e, t in team_keys_needed if kind == "event"]
-    league_team_pairs = [(lg, t) for kind, lg, t in team_keys_needed if kind == "league"]
-    if event_team_pairs:
-        ev_urls = {e for e, t in event_team_pairs}
-        ev_team_ids = {t for e, t in event_team_pairs}
-        for tr in TeamRegistration.query.filter(
-            TeamRegistration.event.in_(ev_urls),
-            TeamRegistration.team.in_(ev_team_ids),
-        ).all():
-            team_pseudonym_by_key[("event", tr.event, tr.team)] = tr.pseudonym
-            team_shortname_by_key[("event", tr.event, tr.team)] = tr.shortname
-    if league_team_pairs:
-        lg_ids = {lg for lg, t in league_team_pairs}
-        lg_team_ids = {t for lg, t in league_team_pairs}
-        for tr in TeamRegistration.query.filter(
-            TeamRegistration.league_id.in_(lg_ids),
-            TeamRegistration.team.in_(lg_team_ids),
-        ).all():
-            team_pseudonym_by_key[("league", tr.league_id, tr.team)] = tr.pseudonym
-            team_shortname_by_key[("league", tr.league_id, tr.team)] = tr.shortname
-
-    def _team_pseudonym(r):
-        if not r.team:
-            return None
-        if r.event:
-            return team_pseudonym_by_key.get(("event", r.event, r.team))
-        if r.league_id:
-            return team_pseudonym_by_key.get(("league", r.league_id, r.team))
-        return None
-
-    def _team_shortname_lookup(r):
-        if not r.team:
-            return None
-        if r.event:
-            return team_shortname_by_key.get(("event", r.event, r.team))
-        if r.league_id:
-            return team_shortname_by_key.get(("league", r.league_id, r.team))
-        return None
-
-    registration_rows = []
-    for r in regs:
-        rcfg = None
-        if r.event:
-            tour = tournaments_by_url.get(r.event)
-            if tour:
-                rcfg = get_registrable_config(tour)
-        elif r.league_id:
-            lg = leagues_by_id.get(r.league_id)
-            if lg:
-                rcfg = lg.registrable_config
-        w = _player_reg_waiver_api(r, rcfg)
-        registration_rows.append(
-            {
-                "event": r.event or (f"league:{r.league_id}" if r.league_id else ""),
-                "team": r.team,
-                "team_pseudonym": _team_pseudonym(r),
-                "team_shortname": _team_shortname_lookup(r),
-                "status": (r.status.value if hasattr(r.status, "value") else str(r.status)),
-                "jersey_name": r.jersey_name,
-                "jersey_number": r.jersey_number,
-                "paid": bool(r.paid),
-                "amount_paid": r.amount_paid or 0.0,
-                "waiver_required": w["waiver_required"],
-                "waiver_status": w["waiver_status"],
-            }
-        )
-
-    return jsonify(
-        {
-            "player": {
-                "id": player.id,
-                "name": player.name,
-                "profile_photo": player.profile_photo,
-                "phone": (player.phone if (current_user.is_authenticated and current_user.id == player_id) else None),
-                "location": player.location,
-                "bio": player.bio,
-            },
-            "registrations": registration_rows,
-            "injuries": [
-                {
-                    "id": inj.id,
-                    "message": inj.message,
-                    "stamp": _dt_iso(inj.stamp),
-                    "active": bool(inj.active),
-                    "show": bool(inj.show),
-                }
-                for inj in injuries
-            ],
-            "player_notes": player_note_rows,
-        }
-    )
-
-
-def _injury_json(inj):
-    return {
-        "id": inj.id,
-        "message": inj.message,
-        "stamp": _dt_iso(inj.stamp),
-        "active": bool(inj.active),
-        "show": bool(inj.show),
-    }
-
-
-@bp.route("/players/<player_id>/injuries/<int:injury_id>", methods=["GET"])
-@login_required
-def get_injury(player_id, injury_id):
-    if current_user.id != player_id:
-        return jsonify({"error": "Forbidden"}), 403
-    injury = Injury.query.filter_by(id=injury_id, player=player_id).first_or_404()
-    return jsonify(
-        {
-            "id": injury.id,
-            "message": injury.message,
-            "stamp": _dt_iso(injury.stamp),
-            "active": bool(injury.active),
-            "show": bool(injury.show),
-        }
-    )
-
-
-@bp.route("/players/<player_id>/injuries", methods=["GET"])
-@login_required
-def list_injuries(player_id):
-    if current_user.id != player_id:
-        return jsonify({"error": "Forbidden"}), 403
-    injuries = Injury.query.filter_by(player=player_id).order_by(Injury.stamp.desc()).all()
-    return jsonify([_injury_json(inj) for inj in injuries])
-
-
-@bp.route("/players/<player_id>/injuries", methods=["POST"])
-@login_required
-def create_injury(player_id):
-    if current_user.id != player_id:
-        return jsonify({"error": "Forbidden"}), 403
-    data = request.get_json() or {}
-    message = (data.get("message") or "").strip()
-    if not message:
-        return jsonify({"error": "message is required"}), 400
-
-    injury = Injury(
-        player=player_id,
-        message=message,
-        show=bool(data.get("show")),
-        active=bool(data.get("active")),
-    )
-
-    custom_date = data.get("custom_date")
-    if custom_date:
-        try:
-            injury.stamp = datetime.strptime(custom_date, "%Y-%m-%d")
-        except ValueError:
-            return jsonify({"error": "Invalid date format. Use YYYY-MM-DD."}), 400
-    db.session.add(injury)
-    db.session.commit()
-    return jsonify(
-        {
-            "id": injury.id,
-            "message": injury.message,
-            "stamp": _dt_iso(injury.stamp),
-            "active": bool(injury.active),
-            "show": bool(injury.show),
-        }
-    )
-
-
-@bp.route("/players/<player_id>/injuries/<int:injury_id>", methods=["PUT"])
-@login_required
-def update_injury(player_id, injury_id):
-    if current_user.id != player_id:
-        return jsonify({"error": "Forbidden"}), 403
-    injury = Injury.query.filter_by(id=injury_id, player=player_id).first_or_404()
-    data = request.get_json() or {}
-    message = data.get("message")
-    if message is not None:
-        message = message.strip()
-        if not message:
-            return jsonify({"error": "message is required"}), 400
-        injury.message = message
-    if "show" in data:
-        injury.show = bool(data.get("show"))
-    if "active" in data:
-        injury.active = bool(data.get("active"))
-    if "custom_date" in data:
-        custom_date = data.get("custom_date")
-        if custom_date:
-            try:
-                injury.stamp = datetime.strptime(custom_date, "%Y-%m-%d")
-            except ValueError:
-                return jsonify({"error": "Invalid date format. Use YYYY-MM-DD."}), 400
-        else:
-            injury.stamp = None
-    db.session.commit()
-    return jsonify(
-        {
-            "id": injury.id,
-            "message": injury.message,
-            "stamp": _dt_iso(injury.stamp),
-            "active": bool(injury.active),
-            "show": bool(injury.show),
-        }
-    )
-
-
-@bp.route("/players/<player_id>/injuries/<int:injury_id>", methods=["DELETE"])
-@login_required
-def delete_injury_api(player_id, injury_id):
-    if current_user.id != player_id:
-        return jsonify({"error": "Forbidden"}), 403
-    injury = Injury.query.filter_by(id=injury_id, player=player_id).first_or_404()
-    db.session.delete(injury)
-    db.session.commit()
-    return jsonify({"ok": True})
-
-
-@bp.route("/teams", methods=["GET"])
-def teams_list():
-    """List teams with optional search."""
-    search = request.args.get("search", "").strip()
-    if search:
-        teams = Team.query.filter(Team.name.contains(search) | Team.id.contains(search)).all()
-    else:
-        teams = Team.query.all()
-    return jsonify(
-        {
-            "teams": [
-                {
-                    "id": t.id,
-                    "name": t.name,
-                    "profile_photo": t.profile_photo,
-                    "location": t.location,
-                }
-                for t in teams
-            ]
-        }
-    )
-
-
-@bp.route("/teams/<team_id>/players", methods=["GET"])
-def team_registration_players(team_id):
-    """Players registered for a team in an event or league (public). Event via query param ?event=.
-    For leagues use ?event=league:league_url."""
-    event_arg = request.args.get("event")
-    if not event_arg:
-        return jsonify({"error": "event required"}), 400
-    team = Team.query.get(team_id)
-    if not team:
-        return jsonify({"error": "Not found"}), 404
-    if event_arg.startswith("league:"):
-        league_id = event_arg[7:]
-        team_reg = TeamRegistration.query.filter_by(
-            team=team_id, league_id=league_id, status=RegistrationStatus.CONFIRMED
-        ).first()
-        if not team_reg:
-            return jsonify({"error": "Not found"}), 404
-        accepted_players = PlayerRegistration.query.filter_by(
-            league_id=league_id, team=team_id, status=RegistrationStatus.CONFIRMED
-        ).all()
-    else:
-        team_reg = TeamRegistration.query.filter_by(
-            team=team_id, event=event_arg, status=RegistrationStatus.CONFIRMED
-        ).first()
-        if not team_reg:
-            return jsonify({"error": "Not found"}), 404
-        accepted_players = PlayerRegistration.query.filter_by(
-            event=event_arg, team=team_id, status=RegistrationStatus.CONFIRMED
-        ).all()
-    players_with_data = []
-    for player_reg in accepted_players:
-        player = Player.query.get(player_reg.player)
-        players_with_data.append(
-            {
-                "registration": {
-                    "player": player_reg.player,
-                    "jersey_name": player_reg.jersey_name,
-                    "jersey_number": player_reg.jersey_number,
-                },
-                "player": (
-                    {
-                        "id": player.id,
-                        "name": player.name,
-                        "profile_photo": player.profile_photo,
-                    }
-                    if player
-                    else None
-                ),
-            }
-        )
-    return jsonify(players_with_data)
-
-
-@bp.route("/teams/<team_id>", methods=["GET"])
-def team_profile(team_id):
-    """Team profile (public)."""
-    team = Team.query.get(team_id)
-    if not team:
-        return jsonify({"error": "Not found"}), 404
-    regs = TeamRegistration.query.filter_by(team=team_id, status=RegistrationStatus.CONFIRMED).all()
-    event_urls = [r.event for r in regs if r.event]
-    tournaments = Tournament.query.filter(Tournament.url.in_(event_urls)).all() if event_urls else []
-    tournament_start = {t.url: t.start_date for t in tournaments}
-
-    tournament_players = {}
-    if current_user.is_authenticated and current_user.id == team_id and is_team(current_user):
-        all_player_ids = set()
-        team_reg_to_players = {}
-        for team_reg in regs:
-            event_key = (
-                team_reg.event if team_reg.event else (f"league:{team_reg.league_id}" if team_reg.league_id else None)
-            )
-            if event_key is None:
-                continue
-            if team_reg.event:
-                accepted_players = PlayerRegistration.query.filter_by(
-                    event=team_reg.event,
-                    team=team_id,
-                    status=RegistrationStatus.CONFIRMED,
-                ).all()
-            else:
-                accepted_players = PlayerRegistration.query.filter_by(
-                    league_id=team_reg.league_id,
-                    team=team_id,
-                    status=RegistrationStatus.CONFIRMED,
-                ).all()
-            team_reg_to_players[event_key] = accepted_players
-            all_player_ids.update(pr.player for pr in accepted_players)
-
-        players_by_id = (
-            {p.id: p for p in Player.query.filter(Player.id.in_(list(all_player_ids))).all()} if all_player_ids else {}
-        )
-
-        for event_key, accepted_players in team_reg_to_players.items():
-            players_with_data = []
-            for player_reg in accepted_players:
-                player = players_by_id.get(player_reg.player)
-                players_with_data.append(
-                    {
-                        "registration": {
-                            "player": player_reg.player,
-                            "jersey_name": player_reg.jersey_name,
-                            "jersey_number": player_reg.jersey_number,
-                        },
-                        "player": (
-                            {
-                                "id": player.id,
-                                "name": player.name,
-                                "profile_photo": player.profile_photo,
-                            }
-                            if player
-                            else None
-                        ),
-                    }
-                )
-            tournament_players[event_key] = players_with_data
-
-    team_notes = []
-    player_played_with_team = False
-    player_tournament_registrations = set()
-    if current_user.is_authenticated and is_player(current_user):
-        player_regs = PlayerRegistration.query.filter_by(
-            player=current_user.id, team=team_id, status=RegistrationStatus.CONFIRMED
-        ).all()
-        player_tournament_registrations = {reg.event for reg in player_regs}
-        player_played_with_team = len(player_tournament_registrations) > 0
-
-    if current_user.is_authenticated:
-        try:
-            candidate_notes = (
-                MatchNote.query.join(Match, Match.uuid == MatchNote.match)
-                .filter(
-                    or_(
-                        and_(MatchNote.target == "team1", Match.team1 == team_id),
-                        and_(MatchNote.target == "team2", Match.team2 == team_id),
-                    )
-                )
-                .order_by(MatchNote.created_at.desc())
-                .all()
-            )
-            match_ids = list({n.match for n in candidate_notes})
-            matches_by_id = (
-                {m.uuid: m for m in Match.query.filter(Match.uuid.in_(match_ids)).all()} if match_ids else {}
-            )
-            match_to_points = {}
-            for n in candidate_notes:
-                m = matches_by_id.get(n.match)
-                if not m:
-                    continue
-
-                can_see_note = False
-                if current_user.id == team_id:
-                    can_see_note = True
-                elif player_played_with_team and current_user.id != team_id:
-                    if m.event in player_tournament_registrations:
-                        can_see_note = True
-                elif is_player(current_user):
-                    if can_head_ref_match(m.event, current_user.id, match=m):
-                        can_see_note = True
-
-                if not can_see_note:
-                    continue
-
-                idx = "-"
-                if n.point_id:
-                    mid = m.uuid
-                    if mid not in match_to_points:
-                        pts = Point.query.filter_by(match=mid).order_by(Point.stamp).all()
-                        match_to_points[mid] = [p.uuid for p in pts]
-                    order = match_to_points.get(mid, [])
-                    if n.point_id in order:
-                        idx = order.index(n.point_id) + 1
-                team_notes.append(
-                    {
-                        "created_at": _dt_iso(n.created_at),
-                        "text": n.text,
-                        "point_index": str(idx),
-                        "match": {
-                            "event": m.event,
-                            "uuid": m.uuid,
-                            "name": m.name,
-                        },
-                    }
-                )
-        except Exception:
-            team_notes = []
-    return jsonify(
-        {
-            "team": {
-                "id": team.id,
-                "name": team.name,
-                "profile_photo": team.profile_photo,
-                "location": team.location,
-                "email": team.email,
-                "website": team.website,
-                "about": team.about,
-            },
-            "registrations": [
-                {
-                    "event": r.event or (f"league:{r.league_id}" if r.league_id else ""),
-                    "pseudonym": r.pseudonym,
-                    "shortname": r.shortname,
-                    "status": (r.status.value if hasattr(r.status, "value") else str(r.status)),
-                    "paid": bool(r.paid),
-                    "amount_paid": r.amount_paid,
-                    "start_date": (_dt_iso(tournament_start.get(r.event)) if r.event else None),
-                }
-                for r in regs
-            ],
-            "team_notes": team_notes,
-            "tournament_players": tournament_players,
-        }
-    )
-
-
 @bp.route("/tournaments/<tournament_url>/matches/<match_id>", methods=["PUT"])
 @login_required
 def update_match_api(tournament_url, match_id):
@@ -4754,187 +4145,6 @@ def list_tags(tournament_url):
         return jsonify({"error": "Forbidden"}), 403
     tags = Tag.query.filter_by(event=tournament_url).order_by(Tag.name).all()
     return jsonify({"tags": [{"id": t.id, "name": t.name, "team": t.team} for t in tags]})
-
-
-@bp.route("/players/<player_id>", methods=["PUT"])
-@login_required
-def update_player_profile(player_id):
-    """Update player profile."""
-    if current_user.id != player_id:
-        return jsonify({"error": "You can only edit your own profile"}), 403
-
-    player = Player.query.get_or_404(player_id)
-    data = request.get_json()
-    if not data:
-        return jsonify({"error": "Invalid JSON"}), 400
-
-    if "name" in data:
-        player.name = data["name"]
-    if "phone" in data:
-        player.phone = data["phone"]
-    if "location" in data:
-        player.location = data["location"]
-    if "bio" in data:
-        player.bio = data["bio"]
-
-    db.session.commit()
-    return jsonify({"success": True})
-
-
-@bp.route("/teams/<team_id>", methods=["PUT"])
-@login_required
-def update_team_profile(team_id):
-    """Update team profile."""
-    if current_user.id != team_id:
-        return jsonify({"error": "You can only edit your own team profile"}), 403
-
-    team = Team.query.get_or_404(team_id)
-    data = request.get_json()
-    if not data:
-        return jsonify({"error": "Invalid JSON"}), 400
-
-    if "name" in data:
-        team.name = data["name"]
-    if "location" in data:
-        team.location = data["location"]
-    if "email" in data:
-        team.email = data["email"]
-    if "website" in data:
-        team.website = data["website"]
-    if "about" in data:
-        team.about = data["about"]
-
-    db.session.commit()
-    return jsonify({"success": True})
-
-
-def _profile_photo_upload_dir():
-    return os.path.join(current_app.root_path, "..", "static", "uploads", "profiles")
-
-
-def _safe_profile_photo_filename(prefix, entity_id):
-    """Sanitize entity id for use in filename (alphanumeric and underscore only)."""
-    safe = "".join(c if c.isalnum() or c == "_" else "_" for c in entity_id)
-    return f"{prefix}_{safe}.jpg"
-
-
-@bp.route("/players/<player_id>/profile-photo", methods=["POST"])
-@login_required
-def upload_player_profile_photo(player_id):
-    """Upload or replace player profile photo. Uses predictable path so overwrites previous."""
-    if current_user.id != player_id or not is_player(current_user):
-        return (
-            jsonify({"error": "You can only upload a photo for your own profile"}),
-            403,
-        )
-    player = Player.query.get_or_404(player_id)
-    data = request.get_data()
-    if not data or len(data) == 0:
-        return jsonify({"error": "No image data"}), 400
-    upload_dir = _profile_photo_upload_dir()
-    os.makedirs(upload_dir, exist_ok=True)
-    # Predictable name: one file per player, always overwritten
-    filename = _safe_profile_photo_filename("player", player_id)
-    file_path = os.path.join(upload_dir, filename)
-    old_path = player.profile_photo
-    rel_path = f"uploads/profiles/{filename}"
-    try:
-        if old_path and old_path != rel_path:
-            old_full = os.path.join(current_app.root_path, "..", "static", old_path)
-            if os.path.isfile(old_full):
-                try:
-                    os.remove(old_full)
-                except OSError:
-                    pass
-        with open(file_path, "wb") as f:
-            f.write(data)
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-    player.profile_photo = rel_path
-    db.session.commit()
-    return jsonify({"success": True, "path": rel_path})
-
-
-@bp.route("/teams/<team_id>/profile-photo", methods=["POST"])
-@login_required
-def upload_team_profile_photo(team_id):
-    """Upload or replace team profile photo. Uses predictable path so overwrites previous."""
-    if current_user.id != team_id or not is_team(current_user):
-        return (
-            jsonify({"error": "You can only upload a photo for your own team profile"}),
-            403,
-        )
-    team = Team.query.get_or_404(team_id)
-    data = request.get_data()
-    if not data or len(data) == 0:
-        return jsonify({"error": "No image data"}), 400
-    upload_dir = _profile_photo_upload_dir()
-    os.makedirs(upload_dir, exist_ok=True)
-    filename = _safe_profile_photo_filename("team", team_id)
-    file_path = os.path.join(upload_dir, filename)
-    old_path = team.profile_photo
-    rel_path = f"uploads/profiles/{filename}"
-    try:
-        if old_path and old_path != rel_path:
-            old_full = os.path.join(current_app.root_path, "..", "static", old_path)
-            if os.path.isfile(old_full):
-                try:
-                    os.remove(old_full)
-                except OSError:
-                    pass
-        with open(file_path, "wb") as f:
-            f.write(data)
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-    team.profile_photo = rel_path
-    db.session.commit()
-    return jsonify({"success": True, "path": rel_path})
-
-
-@bp.route("/players/<player_id>/profile-photo", methods=["DELETE"])
-@login_required
-def delete_player_profile_photo(player_id):
-    """Remove player profile photo."""
-    if current_user.id != player_id or not is_player(current_user):
-        return (
-            jsonify({"error": "You can only remove a photo from your own profile"}),
-            403,
-        )
-    player = Player.query.get_or_404(player_id)
-    old_path = player.profile_photo
-    if old_path:
-        old_full = os.path.join(current_app.root_path, "..", "static", old_path)
-        if os.path.isfile(old_full):
-            try:
-                os.remove(old_full)
-            except OSError:
-                pass
-    player.profile_photo = None
-    db.session.commit()
-    return jsonify({"success": True})
-
-
-@bp.route("/teams/<team_id>/profile-photo", methods=["DELETE"])
-@login_required
-def delete_team_profile_photo(team_id):
-    """Remove team profile photo."""
-    if current_user.id != team_id or not is_team(current_user):
-        return (
-            jsonify({"error": "You can only remove a photo from your own team profile"}),
-            403,
-        )
-    team = Team.query.get_or_404(team_id)
-    old_path = team.profile_photo
-    if old_path:
-        old_full = os.path.join(current_app.root_path, "..", "static", old_path)
-        if os.path.isfile(old_full):
-            try:
-                os.remove(old_full)
-            except OSError:
-                pass
-    team.profile_photo = None
-    db.session.commit()
-    return jsonify({"success": True})
 
 
 @bp.route("/tournaments/<tournament_url>/registrations/player/me", methods=["GET"])
